Added Dynamic File Selection
Added a popup box to allow for dynamic file selection for both data.py and plot.py. Both check to see if it is a valid file (for data.py it has to have "raman" in the name but can't have "normalised" in the name. For plot.py it has to have "_normalised.txt" at the end of the file name. File selection loops if the file is invalid.